PWN-3614 - Separate parsers to separate classes
