+++ app/src/main/java/org/p2p/wallet/App.kt
@@ -21,6 +21,7 @@ import org.p2p.wallet.debug.DebugSettingsModule
 import org.p2p.wallet.debugdrawer.DebugDrawer
 import org.p2p.wallet.feerelayer.FeeRelayerModule
 import org.p2p.wallet.history.HistoryModule
+import org.p2p.wallet.history.HistoryStrategyModule
 import org.p2p.wallet.home.HomeModule
 import org.p2p.wallet.infrastructure.InfrastructureModule
 import org.p2p.wallet.infrastructure.network.NetworkModule
@@ -102,6 +103,7 @@ class App : Application() {
                     SettingsModule.create(),
                     DebugSettingsModule.create(),
                     SwapModule.create(),
+                    HistoryStrategyModule.create()
                 )
             )
         }
