+++ app/src/main/java/org/p2p/wallet/history/repository/remote/TransactionDetailsRpcRepository.kt
@@ -2,18 +2,15 @@ package org.p2p.wallet.history.repository.remote
 
 import org.p2p.solanaj.kits.transaction.TransactionDetails
 import org.p2p.solanaj.kits.transaction.network.ConfirmedTransactionRootResponse
-import org.p2p.solanaj.kits.transaction.parser.OrcaSwapInstructionParser
-import org.p2p.solanaj.kits.transaction.parser.SerumSwapInstructionParser
 import org.p2p.solanaj.model.types.RpcRequest
-import org.p2p.wallet.history.interactor.mapper.SolanaInstructionParser
 import org.p2p.wallet.history.interactor.stream.HistoryStreamItem
+import org.p2p.wallet.history.strategy.TransactionParsingContext
 import org.p2p.wallet.rpc.RpcConstants
 import org.p2p.wallet.rpc.api.RpcHistoryApi
-import org.p2p.wallet.user.interactor.UserInteractor
 
 class TransactionDetailsRpcRepository(
     private val rpcApi: RpcHistoryApi,
-    private val userInteractor: UserInteractor,
+    private val transactionParsingContext: TransactionParsingContext
 ) : TransactionDetailsRemoteRepository {
 
     override suspend fun getTransactions(
@@ -48,32 +45,11 @@ class TransactionDetailsRpcRepository(
     ): List<TransactionDetails> {
         val transactionDetails = mutableListOf<TransactionDetails>()
         transactions.forEach { transaction ->
-            val signature = transaction.transaction?.getTransactionId() ?: return@forEach
-            when {
-                OrcaSwapInstructionParser.isTransactionContainsOrcaSwap(transaction) -> {
-                    val orcaSwapDetails =
-                        OrcaSwapInstructionParser.parse(signature = signature, transactionRoot = transaction)
-                    transactionDetails.add(orcaSwapDetails.getOrThrow())
-                }
-                SerumSwapInstructionParser.isTransactionContainsSerumSwap(transaction) -> {
-                    val serumSwapDetails =
-                        SerumSwapInstructionParser.parse(signature = signature, transactionRoot = transaction)
-                    transactionDetails.add(serumSwapDetails.getOrThrow())
-                }
-                else -> {
-                    transactionDetails.addAll(
-                        SolanaInstructionParser.parse(
-                            signature = signature,
-                            transactionRoot = transaction,
-                            userInteractor = userInteractor,
-                            userPublicKey = tokenPublicKey
-                        )
-                    )
-                }
-            }
+
             transactionDetails.forEach {
                 it.error = transaction.meta.error?.instructionError
             }
+            transactionParsingContext.parseTransaction(transaction)
         }
         return transactionDetails
     }
