+++ app/src/main/java/org/p2p/wallet/history/strategy/TransactionParsingContext.kt
@@ -1,0 +1,12 @@
+package org.p2p.wallet.history.strategy
+
+import org.p2p.solanaj.kits.transaction.network.ConfirmedTransactionRootResponse
+
+interface TransactionParsingContext {
+
+    fun parseTransaction(
+        root: ConfirmedTransactionRootResponse
+    ): ParsingResult
+
+    fun canParse(transaction: ConfirmedTransactionRootResponse): Boolean
+}
