+++ app/src/main/java/org/p2p/wallet/history/strategy/TransactionParsingStrategy.kt
@@ -1,0 +1,15 @@
+package org.p2p.wallet.history.strategy
+
+import org.p2p.solanaj.kits.transaction.TransactionDetailsType
+import org.p2p.solanaj.kits.transaction.network.ConfirmedTransactionRootResponse
+import org.p2p.solanaj.kits.transaction.network.meta.InstructionResponse
+
+interface TransactionParsingStrategy {
+    fun parseTransaction(
+        signature: String,
+        instruction: InstructionResponse,
+        transactionRoot: ConfirmedTransactionRootResponse
+    ): ParsingResult
+
+    fun getType(): TransactionDetailsType
+}
