+++ app/src/test/resources/transactionJson.json
@@ -1,0 +1,119 @@
+{
+  "blockTime": 1617765643,
+  "meta": {
+    "err": null,
+    "fee": 5000,
+    "innerInstructions": [],
+    "logMessages": [
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA invoke [1]",
+      "Program log: Instruction: Transfer",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA consumed 3401 of 200000 compute units",
+      "Program TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA success"
+    ],
+    "postBalances": [
+      34247840,
+      2039280,
+      2039280,
+      1130582400
+    ],
+    "postTokenBalances": [
+      {
+        "accountIndex": 1,
+        "mint": "SRMuApVNdxXokk5GT7XD5cUUgXMBCoAz2LHeuAoKWRt",
+        "uiTokenAmount": {
+          "amount": "0",
+          "decimals": 6,
+          "uiAmount": null,
+          "uiAmountString": "0"
+        }
+      },
+      {
+        "accountIndex": 2,
+        "mint": "SRMuApVNdxXokk5GT7XD5cUUgXMBCoAz2LHeuAoKWRt",
+        "uiTokenAmount": {
+          "amount": "12111",
+          "decimals": 6,
+          "uiAmount": 0.012111,
+          "uiAmountString": "0.012111"
+        }
+      }
+    ],
+    "preBalances": [
+      34252840,
+      2039280,
+      2039280,
+      1130582400
+    ],
+    "preTokenBalances": [
+      {
+        "accountIndex": 1,
+        "mint": "SRMuApVNdxXokk5GT7XD5cUUgXMBCoAz2LHeuAoKWRt",
+        "uiTokenAmount": {
+          "amount": "12111",
+          "decimals": 6,
+          "uiAmount": 0.012111,
+          "uiAmountString": "0.012111"
+        }
+      },
+      {
+        "accountIndex": 2,
+        "mint": "SRMuApVNdxXokk5GT7XD5cUUgXMBCoAz2LHeuAoKWRt",
+        "uiTokenAmount": {
+          "amount": "0",
+          "decimals": 6,
+          "uiAmount": null,
+          "uiAmountString": "0"
+        }
+      }
+    ],
+    "status": {
+      "Ok": null
+    }
+  },
+  "slot": 72550384,
+  "transaction": {
+    "message": {
+      "accountKeys": [
+        {
+          "pubkey": "3h1zGmCwsRJnVk5BuRNMLsPaQu1y2aqXqXDWYCgrp5UG",
+          "signer": true,
+          "writable": true
+        },
+        {
+          "pubkey": "BjUEdE292SLEq9mMeKtY3GXL6wirn7DqJPhrukCqAUua",
+          "signer": false,
+          "writable": true
+        },
+        {
+          "pubkey": "3YuhjsaohzpzEYAsonBQakYDj3VFWimhDn7bci8ERKTh",
+          "signer": false,
+          "writable": true
+        },
+        {
+          "pubkey": "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA",
+          "signer": false,
+          "writable": false
+        }
+      ],
+      "instructions": [
+        {
+          "parsed": {
+            "info": {
+              "amount": "12111",
+              "authority": "3h1zGmCwsRJnVk5BuRNMLsPaQu1y2aqXqXDWYCgrp5UG",
+              "destination": "3YuhjsaohzpzEYAsonBQakYDj3VFWimhDn7bci8ERKTh",
+              "source": "BjUEdE292SLEq9mMeKtY3GXL6wirn7DqJPhrukCqAUua"
+            },
+            "type": "transfer"
+          },
+          "program": "spl-token",
+          "programId": "TokenkegQfeZyiNwAJbNbGKPFXCWuBvf9Ss623VQ5DA"
+        }
+      ],
+      "recentBlockhash": "BPu8bx9VnzeRskapRpXk7u5BAWRuda9t1JEcu71NL8aU"
+    },
+    "signatures": [
+      "rhvqAxG59Q4z4tFP1tJSfkYW1qqkx8A2ttKucjLroSqhS4VEHdrs8VCMb7RZKNq5kXVUqm7F4dLBu8vnM8JJ5q2"
+    ]
+  }
+}
